// packages/genesys-spark-components/src/common-enums.ts
export const enum ListTypeEnum {
  Divider = 'divider',
  Item = 'item',
  Header = 'header'
}

export const enum CalendarModes {
  Single = 'single',
  Range = 'range'
}

export const enum ButtonAccents {
  Primary = 'primary',
  Secondary = 'secondary'
}

export const enum ErrorMessageType {
  Warning = 'warning',
  Error = 'error'
}